+++ src/main/java/com/server/dori/domain/curriculum/presentation/CurriculumController.java
@@ -1,4 +1,4 @@
-package com.server.dori.domain.curriculum.controller;
+package com.server.dori.domain.curriculum.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.server.dori.domain.curriculum.controller.dto.request.CurriculumSurveyRequest;
-import com.server.dori.domain.curriculum.controller.dto.response.CurriculumSurveyResponse;
+import com.server.dori.domain.curriculum.presentation.dto.request.CurriculumSurveyRequest;
+import com.server.dori.domain.curriculum.presentation.dto.response.CurriculumSurveyResponse;
 import com.server.dori.domain.curriculum.service.CurriculumCommandService;
 
 import lombok.RequiredArgsConstructor;
