+++ src/main/java/com/server/dori/domain/curriculum/service/CurriculumCommandService.java
@@ -3,8 +3,8 @@ package com.server.dori.domain.curriculum.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.server.dori.domain.curriculum.controller.dto.request.CurriculumSurveyRequest;
-import com.server.dori.domain.curriculum.controller.dto.response.CurriculumSurveyResponse;
+import com.server.dori.domain.curriculum.presentation.dto.request.CurriculumSurveyRequest;
+import com.server.dori.domain.curriculum.presentation.dto.response.CurriculumSurveyResponse;
 import com.server.dori.domain.curriculum.entity.Curriculum;
 import com.server.dori.domain.curriculum.service.implementation.CurriculumCreator;
